# apps/api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TRIMME API",
    "description": "Public, customer, shop and admin API for the TRIMME salon and barber booking marketplace.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/public/shops/{slug}/professionals": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "Active professionals of an active shop. No contact data.",
        "operationId": "ListPublicShopProfessionals",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicProfessionalResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/professionals": {
      "post": {
        "tags": [
          "Admin: professionals"
        ],
        "summary": "Creates a professional in exactly one shop (audited).",
        "operationId": "AdminCreateProfessional",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfessionalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfessionalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admin: professionals"
        ],
        "summary": "Professionals with shop, status and search filters (paged). Numbers are masked.",
        "operationId": "AdminListProfessionals",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "shopId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProfessionalStatus"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfAdminProfessionalListItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/professionals/{professionalId}": {
      "get": {
        "tags": [
          "Admin: professionals"
        ],
        "summary": "One professional. The WhatsApp number is masked.",
        "operationId": "AdminGetProfessional",
        "parameters": [
          {
            "name": "professionalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfessionalResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin: professionals"
        ],
        "summary": "Edits the profile (audited, optimistic concurrency). The shop cannot change.",
        "operationId": "AdminUpdateProfessional",
        "parameters": [
          {
            "name": "professionalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfessionalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfessionalResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/professionals/{professionalId}/disable": {
      "post": {
        "tags": [
          "Admin: professionals"
        ],
        "summary": "Hides the professional from booking (audited).",
        "operationId": "AdminDisableProfessional",
        "parameters": [
          {
            "name": "professionalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ChangeProfessionalStatusRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfessionalResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/professionals/{professionalId}/enable": {
      "post": {
        "tags": [
          "Admin: professionals"
        ],
        "summary": "Makes a disabled professional active again (audited).",
        "operationId": "AdminEnableProfessional",
        "parameters": [
          {
            "name": "professionalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ChangeProfessionalStatusRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfessionalResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/professionals/{professionalId}/whatsapp": {
      "put": {
        "tags": [
          "Admin: professionals"
        ],
        "summary": "Sets or corrects the WhatsApp number and the notification toggle (audited, number never logged).",
        "operationId": "AdminSetProfessionalWhatsApp",
        "parameters": [
          {
            "name": "professionalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetProfessionalWhatsAppRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfessionalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/professionals/{professionalId}/whatsapp/reveal": {
      "post": {
        "tags": [
          "Admin: professionals"
        ],
        "summary": "Returns the full number once, for a stated reason (audited; never cached).",
        "operationId": "AdminRevealProfessionalWhatsApp",
        "parameters": [
          {
            "name": "professionalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevealProfessionalWhatsAppRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevealedWhatsAppResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/professionals/{professionalId}/avatar": {
      "put": {
        "tags": [
          "Admin: professionals"
        ],
        "summary": "Replaces the photo (JPEG, PNG or WebP, at most 5 MB; audited).",
        "operationId": "AdminUploadProfessionalAvatar",
        "parameters": [
          {
            "name": "professionalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfessionalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin: professionals"
        ],
        "summary": "Removes the photo (audited).",
        "operationId": "AdminRemoveProfessionalAvatar",
        "parameters": [
          {
            "name": "professionalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfessionalResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminProfessionalResponse": {
        "required": [
          "id",
          "shopId",
          "shopSlug",
          "shopNameAr",
          "shopNameEn",
          "slug",
          "nameAr",
          "nameEn",
          "specialtyAr",
          "specialtyEn",
          "bioAr",
          "bioEn",
          "avatarUrl",
          "status",
          "whatsApp",
          "createdAt",
          "updatedAt",
          "version"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "shopId": {
            "type": "string",
            "format": "uuid"
          },
          "shopSlug": {
            "type": "string"
          },
          "shopNameAr": {
            "type": "string"
          },
          "shopNameEn": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "nameAr": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "specialtyAr": {
            "type": [
              "null",
              "string"
            ]
          },
          "specialtyEn": {
            "type": [
              "null",
              "string"
            ]
          },
          "bioAr": {
            "type": [
              "null",
              "string"
            ]
          },
          "bioEn": {
            "type": [
              "null",
              "string"
            ]
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ProfessionalStatus"
          },
          "whatsApp": {
            "$ref": "#/components/schemas/AdminWhatsAppResponse"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "uint32"
          }
        },
        "description": "A professional as the platform admin edits it. Guid AdminProfessionalResponse.ShopId is read-only: it is fixed at creation and no\nrequest can change it (R-NEG-01)."
      },
...
      "ChangeProfessionalStatusRequest": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CreateProfessionalRequest": {
        "required": [
          "shopId",
          "nameAr",
          "nameEn",
          "slug",
          "specialtyAr",
          "specialtyEn",
          "bioAr",
          "bioEn",
          "whatsAppNumber",
          "notificationsEnabled"
        ],
        "type": "object",
        "properties": {
          "shopId": {
            "type": "string",
            "format": "uuid"
          },
          "nameAr": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "slug": {
            "type": [
              "null",
              "string"
            ]
          },
          "specialtyAr": {
            "type": [
              "null",
              "string"
            ]
          },
          "specialtyEn": {
            "type": [
              "null",
              "string"
            ]
          },
          "bioAr": {
            "type": [
              "null",
              "string"
            ]
          },
          "bioEn": {
            "type": [
              "null",
              "string"
            ]
          },
          "whatsAppNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "notificationsEnabled": {
            "type": "boolean"
          }
        },
        "description": "The one shop is chosen here, at creation, and can never be changed afterwards (D-011)."
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "PagedResponseOfAdminProfessionalListItem": {
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminProfessionalListItem"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "The paged envelope every list endpoint returns."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProfessionalStatus": {
        "enum": [
          "Active",
          "Disabled"
        ]
      },
      "PublicProfessionalResponse": {
        "required": [
          "id",
          "slug",
          "nameAr",
          "nameEn",
          "specialtyAr",
          "specialtyEn",
          "bioAr",
          "bioEn",
          "avatarUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "nameAr": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "specialtyAr": {
            "type": [
              "null",
              "string"
            ]
          },
          "specialtyEn": {
            "type": [
              "null",
              "string"
            ]
          },
          "bioAr": {
            "type": [
              "null",
              "string"
            ]
          },
          "bioEn": {
            "type": [
              "null",
              "string"
            ]
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "A professional on the public shop page. Never carries a phone or WhatsApp number (R-PRO-02)."
      },
...
      "RevealedWhatsAppResponse": {
        "required": [
          "number"
        ],
        "type": "object",
        "properties": {
          "number": {
            "type": "string"
          }
        },
        "description": "The full number, returned once per audited reveal. Never cached."
      },
      "RevealProfessionalWhatsAppRequest": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}